refactor: fix some layers violations and init feature-based folder structure (#507)
