+++ apps/container/src/features/calengrade/utils/calendar.ts
@@ -2,7 +2,7 @@ import type { Dayjs } from 'dayjs';
 import dayjs from 'dayjs';
 import * as ics from 'ics';
 
-import type { Calendar } from '../views/Calengrade/types';
+import type { Calendar } from '../types';
 
 const dayOfWeekReturnMappings = {
   Domingo: {
