+++ apps/container/src/features/whatsapp/utils/normalizeTextSearch.ts
