--- apps/container/src/utils/mockedWhatsappGroups.ts
@@ -1,69 +1,0 @@
-// Mock data for preview in WhatsappGroupsView
-export const mockedWhatsappGroups = [
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/GBQropAUsuEGZGXhWYSHrL',
-    codigo: 'BCJ0205-15',
-    campus: 'sa' as const,
-    turma: 'B1',
-    turno: 'noturno',
-    subject: 'Fenômenos Térmicos',
-    teoria: 'Eduardo De Moraes Gregores',
-    pratica: 'Marcos De Abreu Avila',
-  },
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/example2',
-    codigo: 'BCM0506-15',
-    campus: 'sa' as const,
-    turma: 'A2',
-    turno: 'matutino',
-    subject: 'Comunicação e Redes',
-    teoria: 'Maria Silva Santos',
-    pratica: 'João Pedro Oliveira',
-  },
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/example3',
-    codigo: 'BCN0404-15',
-    campus: 'sa' as const,
-    turma: 'C1',
-    turno: 'matutino',
-    subject: 'Geometria Analítica',
-    teoria: 'Ana Carolina Lima',
-    pratica: 'Roberto Carlos Souza',
-  },
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/example4',
-    codigo: 'BCS0001-15',
-    campus: 'sbc' as const,
-    turma: 'B3',
-    turno: 'noturno',
-    subject: 'Base Experimental das Ciências Naturais',
-    teoria: 'Pedro Henrique Costa',
-    pratica: 'Fernanda Rodrigues',
-  },
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/example5',
-    codigo: 'MCM0001-15',
-    campus: 'sa' as const,
-    turma: 'A1',
-    turno: 'matutino',
-    subject: 'Cálculo Diferencial e Integral I',
-    teoria: 'Carlos Eduardo Mendes',
-    pratica: 'Juliana Aparecida Silva',
-  },
-  {
-    season: '2025:2',
-    groupURL: 'https://chat.whatsapp.com/example6',
-    codigo: 'BIR0004-15',
-    campus: 'sa' as const,
-    turma: 'D2',
-    turno: 'noturno',
-    subject: 'Probabilidade e Estatística',
-    teoria: 'Amanda Cristina Alves',
-    pratica: 'Ricardo Monteiro Peixoto',
-  },
-];
